fix(doctor): split SanityField + fix Footer hydration (5 findings, EPIC-7+8)
EPIC-7 only-export-components (4 closed):

  - F-024 SanityField.tsx:25:37: moved buildStudioEditHref +
    buildSanityDataAttribute to sanity-field.lib.ts and the
    three type exports to sanity-field.types.ts. SanityField.tsx
    now exports only the component. Barrel src/components/cms/
    index.ts updated to re-export from all three files.

  - F-025 src/app/opengraph-image.tsx:4:14: Next.js App Router
    metadata file convention requires runtime + size +
    contentType named exports alongside the default image
    component. Disabled the rule file-wide with a header
    comment explaining the Next.js convention.

  - F-026 src/app/twitter-image.tsx:4:14: same pattern.

EPIC-8 rendering-hydration-mismatch-time (1 closed):

  - F-038 Footer.tsx:62: 'new Date().getFullYear()' was called
    in render. Hoisted the computation to a module-level
    constant (CURRENT_YEAR) so SSR and client see the same
    value for any given build. The component now seeds
    useState with CURRENT_YEAR and only re-derives in a
    useEffect if the year changes during a long-lived tab.

Verified:
- npm run doctor: 0 errors, 0 warnings ('No issues found!')
- vitest: 29/29 test files, 266/266 tests pass
- SanityField import paths updated; barrel re-exports all
  three modules
